feat(translation): Separate KebabMenu component and optimize code
