+++ packages/web/app/components/Sidebar/Sidebar.jsx
@@ -1,10 +1,9 @@
 import React, { useCallback, useState } from 'react';
 import styled from 'styled-components';
-import { Box, Button, Divider, IconButton, List, Typography, Menu } from '@material-ui/core';
-import { Launch, NavigateBefore, NavigateNext, MoreVert } from '@material-ui/icons';
+import { Box, Button, Divider, IconButton, List, Typography } from '@material-ui/core';
+import { NavigateBefore, NavigateNext } from '@material-ui/icons';
 import { useDispatch, useSelector } from 'react-redux';
 import { push } from 'connected-react-router';
-import { keyBy, mapValues } from 'lodash';
 import { getCurrentRoute } from '../../store/router';
 
 import { LogoLight, LogoLightNoText } from '../Logo';
@@ -17,10 +16,7 @@ import { checkAbility } from '../../utils/ability';
 import { useAuth } from '../../contexts/Auth';
 import { useApi } from '../../api';
 import { TranslatedText } from '../Translation/TranslatedText';
-import { useLocalisation } from '../../contexts/Localisation';
-import { ChangeLanguageModal } from '../ChangeLanguageModal';
-import { LanguageSelector } from '../LanguageSelector';
-import { useTranslationLanguages } from '../../api/queries';
+import { KebabMenu } from './KebabMenu';
 
 const Container = styled.div`
   display: flex;
@@ -137,47 +133,8 @@ const LogoutButton = styled(Button)`
   padding-right: 0;
 `;
 
-const SupportDesktopLink = styled.a`
-  margin-top: 4px;
-  font-weight: 400;
-  font-size: 11px;
-  line-height: 15px;
-  text-decoration: underline;
-  color: ${Colors.white};
-`;
-
-const KebabMenuItem = styled.div`
-  font-weight: 400;
-  font-size: 11px;
-  line-height: 15px;
-  color: ${Colors.white};
-  padding: 4px;
-  cursor: pointer;
-  border-radius: 4px;
-  :hover {
-    background: rgba(255, 255, 255, 0.15);
-  }
-`;
-
-const StyledIconButton = styled(IconButton)`
-  color: white;
-  align-self: start;
-  padding: 12px 0 0 0;
-`;
-
 const StyledMetadataBox = styled(Box)`
   margin-bottom: 5px;
-`;
-
-const StyledMenu = styled(Menu)`
-  & .MuiPaper-root {
-    border: 1px solid ${Colors.outline};
-    background: ${Colors.primaryDark};
-    width: 124px;
-  }
-  & .MuiList-padding {
-    padding: 4px;
-  }
 `;
 
 const getInitials = string =>
@@ -205,28 +162,13 @@ const isHighlighted = (currentPath, menuItemPath, sectionIsOpen, isRetracted) =>
 };
 
 export const Sidebar = React.memo(({ items }) => {
-  const [anchorEl, setAnchorEl] = useState(null);
-  const [isChangingLanguage, setChangingLanguage] = useState(false);
-  const open = Boolean(anchorEl);
   const [selectedParentItem, setSelectedParentItem] = useState('');
   const [isRetracted, setIsRetracted] = useState(false);
   const { agentVersion } = useApi();
   const { facility, centralHost, currentUser, onLogout, currentRole } = useAuth();
   const currentPath = useSelector(getCurrentRoute);
   const dispatch = useDispatch();
-  const { getLocalisation } = useLocalisation();
   const extendSidebar = () => setIsRetracted(false);
-
-  const { data = {} } = useTranslationLanguages();
-  const { languageNames = [], languagesInDb = [] } = data;
-  const languageDisplayNames = mapValues(keyBy(languageNames, 'language'), 'text');
-
-  const languageOptions = languagesInDb.map(({ language }) => {
-    return {
-      label: languageDisplayNames[language],
-      value: language,
-    };
-  });
 
   const onPathChanged = newPath => dispatch(push(newPath));
 
@@ -241,21 +183,12 @@ export const Sidebar = React.memo(({ items }) => {
     }
   };
 
-  const onOpenKebabMenu = (event) => {
-    setAnchorEl(event.currentTarget);
-  };
-
-  const handleCloseKebabMenu = () => {
-    setAnchorEl(null);
-  };
-
   const handleRetractButtonClick = useCallback(() => setIsRetracted(true), []);
 
   const handleExtendButtonClick = useCallback(extendSidebar, []);
 
   const initials = getInitials(currentUser.displayName);
   const roleName = currentRole?.name ?? currentUser?.role;
-  const supportUrl = getLocalisation('supportDeskUrl');
 
   return (
     <Container $retracted={isRetracted}>
@@ -350,48 +283,7 @@ export const Sidebar = React.memo(({ items }) => {
               </Box>
             </StyledUserInfoContent>
           )}
-          <StyledIconButton
-            onClick={onOpenKebabMenu}
-          >
-            <MoreVert />
-          </StyledIconButton>
-          <StyledMenu
-            anchorEl={anchorEl}
-            keepMounted
-            open={open}
-            onClose={handleCloseKebabMenu}
-            anchorOrigin={{
-              vertical: 'top',
-              horizontal: 'left',
-            }}
-            transformOrigin={{
-              vertical: 'top',
-              horizontal: 'right',
-            }}
-          >
-            {/* If multiple languages not implemented, no need for the modal to show */}
-            {languageOptions.length > 1 && <KebabMenuItem
-              onClick={() => setChangingLanguage(true)}
-            >
-              <TranslatedText stringId="general.language.change" fallback="Change language" />
-            </KebabMenuItem>}
-            <KebabMenuItem>
-              <SupportDesktopLink href={supportUrl} target="_blank" rel="noreferrer">
-                <TranslatedText
-                  stringId="sidebar.externalLink.supportCentre"
-                  fallback="Support centre"
-                />
-                <Launch style={{ marginLeft: '5px', fontSize: '12px' }} />
-              </SupportDesktopLink>
-            </KebabMenuItem>
-            <ChangeLanguageModal
-              maxWidth='lg'
-              fullWidth={false}
-              open={isChangingLanguage}
-              onClose={() => setChangingLanguage(false)}
-            />
-          </StyledMenu>
-          <LanguageSelector />
+          <KebabMenu />
         </UserInfo>
         {!isRetracted && (
           <>
